Fix broken Room code generation
We use Kotlin CodeGen for Room in the app module, but not in other modules. This along with KSP2 breaks the update to Kotlin 2.2.0

We can turn KSP2 off for now and ensure that any module that adds the Room compiler generates Kotlin code.

This will be the default in Kotlin 2.2.0+ but we cannot use it right now as we get occasional build issues. See the Asana task for more info.