keep input text box one line until interaction (#6468)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1210835160047735?focus=true

### Description
If the initial Input Screen content carried over from the omnibar is a
URL, keep the Input Screen's text box a single line high unless clicked.

The idea is that the `EditText` is initialized as one line, which
requires it to be scrollable horizontally. This means that operations on
URL initiated by a long-click (context menus for copy/cut, and selection
management) can be done within the horizontally scrolling box. This
behavior exactly matches production omnibar.

However, as soon the input box is clicked, it expands to a maximum of 8
lines.

I also cleaned up unused now transition params.

### Steps to test this PR

- [x] Open input screen from a website with long-ish URL, that wouldn't
normally fit in one line.
- [x] Verify input box continue to be one line.
- [x] Verify you can long click without expanding the box.
- [x] Verify clicking on the input box expands to fit all of the content
(or reaches max of 8 lines).
- [x] Verify the cursor remains where you clicked.

Example:


https://github.com/user-attachments/assets/42ad7047-f26d-4db3-aa64-33c045051be0


### Known issues
Sometimes, the text doesn't immediately fill the available space after
expansion,
[example](https://app.asana.com/1/137249556945/task/1210835160047735/comment/1210890848594073?focus=true).

---------

Co-authored-by: joshliebe <joshliebe@gmail.com>