Avoid crash on onQuickAccessListChanged (#6509)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/488551667048375/task/1210908558625303?focus=true

### Description

`onQuickAccessListChanged` always assumed that `favourites` size is
greater than `newList`. However from the crash logs we've seen tells us
that that is not the case. The current solution results to
f`avourites.takeLast` being called with a negative value.

To fix this, we only add items to the `newlist` if the `favourites` size
is actually bigger.

### Steps to test this PR
Smoke test dragging favourites in the new tab screen