Launch device auth settings differently for TCL devices (#6540)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1200930669568058/task/1210133000985922?focus=true

### Description
Changes the way the system activity for setting up device auth is
launched on `TCL`-manufactured devices.
- Observed them behaving strangely in that they would not show the
system device auth activity
- In the logs, noticed them stating `no hardware available` 
- After investigating, concluded they are mishandling the provided flags
around biometrics capabilities and not handling the `or` flag properly

We don't require biometrics to be in place; only `credentials`. So for
TCL devices as a workaround will provide the intent extra to indicate we
only need `credentials`.

### Steps to test this PR

- [x] Disable device auth and visit password management (`Overflow ->
Passwords`)
- [x] Tap `Set up in Settings` button
- [x] Verify you see the system settings to set up
pin/password/fingerprint etc...
- [ ] You won't have a TCL device so will need to just QA-optional that
change; i've tested it with a real device.

Co-authored-by: Craig Russell <1336281+CDRussell@users.noreply.github.com>