Remove old password management list view implementation (#6473)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/608920331025315/task/1209443099736785?focus=true

### Description
In https://github.com/duckduckgo/Android/pull/5654 we changed the
implementation of the password management screen so that everything was
rendered as a list item. We kept the old implementation around as a
fallback, guarded by a kill switch. There's been no reported issues with
the new approach.

This PR removes the old implementation. 

### Steps to test this PR

#### Normal Autofill
- [x] Smoke test normal Autofill screens, including password management
and password settings screens
- [x] Smoke test the `AutofillService` screen that happens when you to
choose to open the passwords screen to pick your own one to autofill.

---------

Co-authored-by: Craig Russell <1336281+CDRussell@users.noreply.github.com>