Avoid usage of requireActivity in download callback (#6558)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/488551667048375/task/1211022417452920?focus=true

### Description
After downloading a file we have a callback which shows a snackbar;
there is button on it to open the file. From a stack trace it seems
possible the activity is gone and the use of `requireActivity()` results
in a crash. Switching to avoiding the crash and doing nothing if the
`activity` is null when we try.

Fixes https://github.com/duckduckgo/Android/issues/6557

### Steps to test this PR
- [ ] QA optional
- [ ] If you want, download a file (e.g.,
https://file-examples.com/index.php/sample-images-download/sample-jpg-download/)
and tap on the snackbar to view it once done

Co-authored-by: Craig Russell <1336281+CDRussell@users.noreply.github.com>