do not open the Input Screen when user pref toggled on and immediately off (#6615)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1211091584703241?focus=true

### Description
Fixes an issue where toggling the experimental address bar on and then
immediately off again in "AI Features" settings still caused the Input
Screen to open when returning to the browser (if left on a New Tab
Page).

#### Root Cause
The Input Screen preferences flow in the `BrowserTabViewModel` (and its
nested flow that launches the Input Screen on NTP) was being restarted
every time the preference changed. As a result, toggling the setting
caused a re-evaluation, which queued a command to launch the Input
Screen when returning to the browser.

#### Fix
Instead of observing the Input Screen preference flow directly, we now
filter out events when the preference is disabled. This prevents
unnecessary re-evaluations on every toggle, while still ensuring that
whenever the tab changes, the current preference state is correctly
respected.

### Steps to test this PR

- [x] Install a clean build of the app.
- [x] Open a new tab and ensure that the address bar gets focused but
the Input Screen doesn't open.
- [x] Go to Settings -> AI Features
- [x] Enabled Experimental Address Bar
- [x] Disabled Experimental Address Bar
- [x] Go back to the browser and verify that Input Screen is not opened
- [x] (optional) Try some additional test steps from
https://github.com/duckduckgo/Android/pull/6604 to ensure that the
feature still works as expected otherwise