remove singleOmnibarFeature feature flag  (#6658)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1211098896784346?focus=true

### Description

Removes `singleOmnibarFeature` feature flag and cleans up associated
code paths and resources that are now unused.

### Steps to test this PR

An easiest (potentially) way to test this PR is to run two devices (or
emulator instances) side-by-side - one with a build from
https://github.com/duckduckgo/Android/commit/05c58cca28278d4ff232f534ebc5a3400b0629b3,
right before this PR, the other from the top of the PR branch.

Perform the steps together on both devices and look for any differences
in presentation, behavior, colors, or anything else. There should be
none.

- [x] Clean install the apps.
- [x] Go through onboarding, verify that privacy shield and fire button
highlighting works.
- [x] Interact with the omnibar buttons, ensure they are all responsive.
- [x] Disable Duck.ai in the address bar (Settings -> AI Features ->
Duck.ai Shortcuts) and verify the Duck.ai icon is gone from the omnibar.
- [x] Enable voice input and verify the microphone appears in the
omnibar.
- [x] Re-enable Duck.ai in the address bar and verify that both icons
are visible.
- [x] When focusing on the address bar, verify that the view expands and
buttons re-position in an animated way.
- [x] Enable the light/dark theme and check the colors for the omnibar
background, setting pages background, and toolbars match on both
versions.
- [x] Go to Settings -> Developer Settings -> Custom Tabs, set the
browser as default and open a custom tab. Verify that it looks identical
on both devices.
- [x] Go to Settings -> AI Features and enable the experimental address
bar. Verify that clicking on the omnibar or creating a new tab opens the
new Input Screen.
- [x] Consider any other scenarios that might be worth testing.