Remove showOmnibar from navigationStateChanged to fix easter egg logos
This is a leftover from Josh's loading bar experiment. Ondrej needed to account for that when adding the new Bottom Omnibar here but as Josh has removed the experiment it is no longer needed. I tested and it makes no difference and now acts the same way as the top omnibar