Fix url detector intent detection logic (#6687)
Task/Issue URL: https://app.asana.com/1/137249556945/project/1204912272578138/task/1211178371917515?focus=true

### Description
When the user enters URLs like `chrome://...` and/or `edge://...` or similar, our logic fails to properly handle them thinking they are app intents. The result is that we don't navigate anywhere, yet we update the address bar.

The correct behavior (like in eg. DDG MacOS browser) should be to perform a search.

### Steps to test this PR

_Test_
- [x] build and install from this branch
- [x] open the browser and open a tab
- [x] navigate to chrome://flags
- [x] verify the browser should perform a search
- [x] navigate to edge://flags
- [x] verify the browser should perform a search

Repeating the test above in develop results in the browser navigating anywhere which is not the intended behavior