show Duck.ai fragment immediately if opened from outside of the BrowserActivity (#6683)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1211037269945827?focus=true

### Description

Adds a check whether Duck.ai should animate in or be opened immediately.
The former is only allowed if the command to open Duck.ai fragment
originated while the `BrowserActivity` was already visible and can
smoothly transition to the Duck.ai fragment. Otherwise, the fragment is
show immediately to avoid janky transitions.

### Steps to test this PR
- [x] With the new Input Screen disabled, go to a website and click
Duck.ai in the omnibar. Go back, and click the button again. Verify that
Duck.ai opens with a transition.
- [x] Click the address bar, type in some text, and click the Duck.ai
icon. Verify that it opens with a transition.
- [x] Go to Settings -> AI features and enabled the Input Screen.
- [x] Go back to the website and click the Duck.ai button. Verify
Duck.ai opens with a transition.
- [x] Click the omnibar, switch to Duck.ai, type in and submit a prompt.
Verify that Duck.ai opens immediately without a transition.
- [x] Go back and verify that the fragment closes with a transition back
to the website.

_before_


https://github.com/user-attachments/assets/cdfdee2a-68cc-4377-b36d-e349a240cf0a

_after_


https://github.com/user-attachments/assets/2abcf956-400f-4568-97a4-7f22e9a8b1b2