update Duck.ai icon (omnibar) click behavior (#6724)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1210888855345532?focus=true

### Description
Simplifies behavior of the Duck.ai button in the omnibar.

If the button is clicked while the omnibar has focus, start a new chat
and auto-prompt Duck.ai with contents of the omnibar. If the buttons is
clicked withe the omnibar doesn't have focus, open existing chat if
within 60 minutes of starting it, otherwise, open a new empty chat.

The only exception is New Tab Page which still opens the existing chat
(if within 60 minutes since starting it) when omnibar is focused and
empty.

### Steps to test this PR

- [x] Open Duck.ai and start a new chat.
- [ ] Go to New Tab Page.
- [x] While omnibar bar **is** focused, click Duck.ai button and verify
that the previous chat is opened.
- [x] While omnibar **is not** focused, click Duck.ai button and verify
that the previous chat is opened.
- [x] Type anything into the omnibar and while omnibar bar **is**
focused, click Duck.ai button and verify that a new chat is opened with
auto-prompt.
- [x] Go to SERP.
- [x] While omnibar bar **is** focused, click Duck.ai button and verify
that a new chat is opened with auto-prompt.
- [x] While omnibar bar **is not** focused, click Duck.ai button and
verify that the previous chat is opened.
- [x] Go to a website.
- [x] While omnibar bar **is** focused, click Duck.ai button and verify
that a new chat is opened with auto-prompt.
- [x] While omnibar bar **is not** focused, click Duck.ai button and
verify that the previous chat is opened.
- [x] (optional) Change device's time to > 1 hour into the future.
- [x] While omnibar bar **is not** focused, click Duck.ai button and
verify that a new empty chat is opened.