Fix: Missing menu button (#6756)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1207418217763355/task/1211192395246455?focus=true

### Description

This PR fixes a but when the `...` toolbar button in tab switcher is not
displayed when the app is run in a clean state. The reason is that the
privacy config is not downloaded yet and the required feature flag is
disabled by default.

The fix includes cleaning up the relevant feature flag, which has been
already fully enabled a long time ago.

### Steps to test this PR

_Tab switcher bug fix_
- [x] Clean the app storage
- [x] Run the app and quickly skip the onboarding
- [x] Open the tab switcher
- [x] Verify the `...` toolbar button is visible

_Multi-selection_
- [x] In the tab switcher, enable the selection mode
- [x] Smoke test multi-selection and check that everything works as
expected