Alternative approach for Input Screen landscape dimensions calculation (#6744)
This is an alternative approach for
https://github.com/duckduckgo/Android/pull/6740 which uses dimensions
instead of programmatically checking for orientation.

According to [window size
classes](https://developer.android.com/develop/ui/compose/layouts/adaptive/use-window-size-classes),
width over 600dp is nearly guaranteed to be in landscape, so we can use
a dimension class to change the margins only then.

---------

Co-authored-by: joshliebe <joshliebe@gmail.com>