Subscription webview go back action (#6752)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/488551667048375/task/1211073283656160?focus=true

### Description
This PR changes the way navigating back works letting the FE decide in
some cases when it can and cannot happen

### Steps to test this PR

- [x] Apply the staging patch and VPN into US
- [x] Go to https://duckduckgo.com/pro?featurePage=stripe
- [x] Complete the purchase using the test cards for Stripe
- [ ] On the welcome screen, the icon should be a cross rather than an
arrow.
- [ ] Pressing either the cross or the back button should take you back
to settings (doing this on develop would take you back to Stripe)
- [ ] Now that you are subscribe go to settings and click on "Need help
with your subscription"
- [ ] On that webview navigate to some links, see how the cross turns
into an arrow and if you now press it or use the back button you
navigate back rather than close the screen.