try-catch startForeground inside notifyVpnStart (#6833)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1202552961248957/task/1211418384396819?focus=true

### Description
In https://github.com/duckduckgo/Android/pull/6808 we bubbled up the
`ForegroundServiceStartNotAllowedException` to notifyVpnStart function,
when calling `startForeground` method.

This didn't get rid of the crashes but was a bit of an improvement.

This PR will try-catch the startForeground call and return the
appropriate error so that the error path can be executed. The error path
sends a pixel (m_vpn_ev_notify_start_failed_c/d) are expected to
increase and will also stop the service showing the disabled
notification. Which is a better UX

### Steps to test this PR

Smoke test the VPN and AppTP