Use ktlint 0.50.0
Looks like we can go back to the last ktlint release before the big rule changes. Any earlier and I saw a lot of violations with "UNDEFINED" for what they were and line numbers etc. I don't know if this will break when we update to Kotlin 2.x.

I'm defining an .editorconfig path to ensure that we're always using our root one whenever this might be added to other modules.

I've added the kdoc rule to be disabled. It stops warnings for things like this:

aProperty = { /* kdoc */ }