Enhance address bar trackers animation with new scenes and improved logic
This moves the former experimental trackers animation to it's new state where it mimics the Cookies animation by sliding out but retains it's count up behaviour from the original. We moved to this style to maintain consistency with the animations here

We added some new behaviour in the TrackerCountAnimator which means we only animate the count up if we have greater than 4 trackers, otherwise we wait and the end the animation.

We need to expose getTrackerAnimationStartCount so that it's possible to set the correct widths for the trackers count TextViews

We'll tidy up the BrowserLottieTrackersAnimatorHelper in a later commit but I wanted to leave this as is so you can see the previous experiment behaviour