Cache autoconsent settings in singleton (#6910)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1163321984198618/task/1211608075081851?focus=true

### Description
We are currently parsing the autoconsent settings JSON object every time
the `init` message is received (every page load). As this is a heavy
operation (parsing a >300kb string), and the output changes infrequently
(config changes max once per day), we can cache the value when the
config is downloaded, then only reparse if the settings value changes.

### Steps to test this PR

1. Visit
https://privacy-test-pages.site/features/autoconsent/index.html.
 2. The reject button should be clicked.