Filter CPM rules in native before sending to JS (#6922)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1163321984198618/task/1211626409836529?focus=true

### Description
Filters out non-relevant site-specific rules before constructing a JSON
message to send to the autoconsent script.

- Reduces the JSON message size for the latest CPM ruleset from >700kb
to ~50kb (note, we currently using a smaller list on Android, because
this is causing crashes).
- We only partially define the compact rule type, because moshi isn't
able to properly parse tuples. This means we have to cast specific
values to use them. This should be safe, because the compact rules are
typed at build time.

### Steps to test this PR

1. Visit
https://privacy-test-pages.site/features/autoconsent/index.html.
2. The reject button should be clicked.


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1211626409836529

---------

Co-authored-by: joshliebe <joshliebe@gmail.com>