Refactor: Make Shapes and Typography default theme parameters
This change simplifies the `DuckDuckGoTheme` by providing default values for `shapes` and `typography` directly in the `ProvideDuckDuckGoTheme` function signature.

The `Shapes` and `Typography` are now instantiated as top-level properties, making them readily available as defaults and removing the need for manual creation within the `DuckDuckGoTheme` composable. Documentation links to Figma have also been added for colors, shapes, and typography.