Measure concurrent load requests (#6190)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/414730916066338/task/1211393669911397?focus=true

### Description

The PR adds a counter that measures all concurrent load requests across
all `BrowserWebViewClient` instances. It uses `AtomicInteger` and
`ConcurrentHashMap` to track the requests in progress and the count is
reported to the `PageLoadedHandler`.

For the counter, there's a 30s time-out (maybe it should be longer?) so
that we don't count requests that never finish.

### Steps to test this PR

_DB migration_
- [x] Build the app from `develop` and open it
- [x] Open a couple of tabs using these sites: "bbc.com", "weather.com",
"reddit.com" or "ebay.com"
- [x] Wait a bit for all of them to load
- [x] Rebuild the app from this PR branch and run it
- [x] Notice the app doesn't crash and runs normally
- [x] Open the database inspector (app database,
`page_loaded_pixel_entity` table) and check that
`isTabInForegroundOnFinish` is false, `activeRequestsOnLoadStart` is 0
and `concurrentRequestsOnFinish` is 0 for all requests in the table

_1 request_
- [x] Kill the app
- [x] Start the app
- [x] Open a new tab and load `wikipedia.org`
- [ ] Open the database inspector and check that
`isTabInForegroundOnFinish` is false, `activeRequestsOnLoadStart` is 0
and `concurrentRequestsOnFinish` is 0 for the request

_Multiple requests_
- [x] Keeping the same app instance, swipe fast a couple of times
between tabs
- [x] Notice that the concurrent tab requests are reported
- [x] Open the database inspector and check that
`isTabInForegroundOnFinish` is true for some requests and
`activeRequestsOnLoadStart` and `concurrentRequestsOnFinish` is not 0
for some