add `search_suggestion_index` pixel param to search suggestion click pixels (#6979)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1211716661936840?focus=true

### Description
Adds a parameters to search suggestion click pixels with the index of
the clicked suggestion. The index ignores history, favorites, bookmarks,
and other suggestion types which are not strictly a query suggestion or
a website suggestion.

For example, if the suggestions list includes:
- (history) hey
- (favorite) heeey.com/wave
- hey.com
- hey
- hello
- hello there

and users clicks on 'hello', the `search_suggestion_index=2` param will
be appended.

### Steps to test this PR

- [x] Open the app, focus on the omnibar and type in a query.
- [x] Verify that clicking on a search suggestion sends an
`m_autocomplete_click_phrase` pixel with `search_suggestion_index`
matching the position of the suggestion in the list.
- [x] Verify that clicking on a website search suggestion sends an
`m_autocomplete_click_website` pixel with `search_suggestion_index`
matching the position of the suggestion in the list.