Split Omnibar: Omnibar layout refactoring (#6983)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1207418217763355/task/1211565479656354

### Description

This PR refactors the `OmnibarLayout` and `SingleOmnibarLayout`.
`OmnibarLayout` was renamed to `LegacyOmnibarLayout` and a new
`OmnibarLayout` is a unified, merged version of `OmnibarLayout` and
`SingleOmnibarLayout`.

Thew new omnibar is guarded by a `useUnifiedOmnibarLayout` feature flag.
To make that work, an `OminbarView` interface needed to be extracted,
which is now used within the `Omnibar` class. Depending on the flag
value, either the old `LegacyOmnibarLayout ` or the new `OmnibarLayout`
is used. The feature flag is enabled by default.

Also, some of the interfaces from `OmnibarLayout` had to be extracted
(`StateChange`, `Decoration`) to avoid duplication.

### Steps to test this PR

_FF enabled_
- [x] Verify that onboarding highlighting works
- [x] Verify all buttons work
- [x] Verify tab swiping works
- [x] Verify tracker animations work
- [x] Verify the cookies animations work
- [x] Verify that omnibar hiding on scroll works

_FF disabled_
- [x] Disable the `useUnifiedOmnibarLayout` FF in the internal settings
- [x] Verify that onboarding highlighting works
- [x] Verify all buttons work
- [x] Verify tab swiping works
- [x] Verify tracker animations work
- [x] Verify the cookies animations work
- [x] Verify that omnibar hiding on scroll works