SERP Settings Sync: Remove KBG flag if feature enabled (#7014)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/72649045549333/task/1211702809875729?focus=true

### Description

Added a conditional check for the `serpSettingsSync` feature flag before
applying the DuckAI SERP hiding logic to ensure when we complete the
rest of the work we're getting the final behaviour we expect and the kbg
param is not inteferring with anyhing.
  
`serpSettingsSync` is now off by default so we do not mess with kbg for
internal users until the rest of the work is complete.

### Steps to test this PR

_KBG Param_

- [x] Ensure Duck.ai is enabled in native settings
- [x] Perform a search e.g. FPL
- [x] Check logs that the search url contains `kbg=-1`​
- [x] Enable `serpSettingsSync`​ feature flag (off by default for now)
- [x] Perform a search e.g. FPL
- [x] Check logs that the search url does **not** contain `kbg=-1`​

### UI changes

N/A