Ensure FF check happens on io for 1st party cookie expiry work scheduling (#7081)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/488551667048375/task/1211876763499720?focus=true

### Description
Moves the feature flag checks to `io` instead of `main`. This is
triggered from `MainProcessLifecycleObserver` so is an important hot
path for app launch / ANR prevention.

### Steps to test this PR
- QA optional
- [ ] [optional]: add a log statement to
`FirstPartyCookiesModifierWorker.doWork()` and make sure it continues to
fire with this change in place when the process is created
- [ ] [optional]: verify the `StrictMode` violation warning no longer
shows on app process creation

Co-authored-by: Craig Russell <1336281+CDRussell@users.noreply.github.com>