Create new feature flag for handling about blank web pages (#7132)
Task/Issue URL: 

https://app.asana.com/1/137249556945/project/1211724162604201/task/1211850430148973?focus=true

### Description

Added a new toggle `handleAboutBlank()` to the
`AndroidBrowserConfigFeature` interface. This toggle is internally
always enabled and defaults to `false`. It will be used to determine
whether the app should handle "about:blank" URLs based on remote
configuration.

### Steps to test this PR

_Handle About Blank Feature_
- [x] Verify that the new toggle is properly annotated with
`@Toggle.InternalAlwaysEnabled`
- [x] Verify that the toggle defaults to `false` when remote config is
not present

### UI changes
No UI or functional changes in this PR.