Add surfaces parameter to RemoteMessage and updated related classes. (#7063)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1200581511062568/task/1211783737656573?focus=true

### Description
Added support in the RemoteMessage model for displaying remote messages
on different surfaces.
This PR introduces a new `Surface` enum with two values: `MODAL` and
`NEW_TAB_PAGE` and a new `surfaces` field in the RemoteMessage.

If no surfaces are specified in the JSON, messages will default to
appearing on the `NEW_TAB_PAGE` surface.

### Steps to test this PR

- [ ] Checkout this branch.
- [ ] Ensure you are using a custom JSON RMF config. See details in the
"Useful for testing" section in the Asana Task
https://app.asana.com/1/137249556945/project/1200581511062568/task/1211783737656573?focus=true.
- [ ] Verify that remote messages are parsed and stored properly.

### NO UI changes