Make "bookmark added" dialog specific to that use case (#7099)
Task/Issue URL: https://app.asana.com/1/137249556945/project/1208273769335188/task/1211910160515992?focus=true

### Description

This is taking `BookmarksBottomSheetDialog`, which is generic but only ever used for a specific use case (bookmark added confirmation) and just making it specific to that use case. This simplifies it as well as better preparing it for extension to support adding an option to let users know they can enable sync.

- `BookmarksBottomSheetDialog` --> `BookmarkAddedConfirmationDialog`

Strings moving to the correct module is done in https://github.com/duckduckgo/Android/pull/7127 to de-noise this one..

### Steps to test this PR

**Auto-dismisses**

- [x] Visit site, tap overflow and choose `Add Bookmark`.
- [x] Verify the bookmark added dialog shows and looks the same as prod does
- [x] Don't interact with it; verify it auto-dismisses (after ~3.5s)
- [x] Choose `Edit Bookmark` and delete it

**Edit (not as favorite)**

- [x] Choose `Add Bookmark` again, don't toggle the `Add to Favorites` but do tap on `Edit Bookmark`. Verify it isn't marked as favorite.
- [x] Delete it

**Edit (when a favorite)**

- [x] Choose `Add Bookmark` again, this time toggle the `Add to Favorites` to enabled
- [x] Tap `Edit Bookmark`. Verify it is still marked as a favorite.