Remove the siteErrorHandlerKillSwitch feature flag (#7157)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/488551667048375/task/1212002086443929?focus=true

### Description

This PR removes the `siteErrorHandlerKillSwitch` feature flag. 

One thing I added is that the error handling code is now run on the IO
thread in a coroutine to minimize blocking of the main thread when
setting the `site` value.

### Steps to test this PR

- Open `amazon.com` (either from SERP or manually)
- Type in **manually** `amazon.com` and load the page multiple times.
- Verify that address bar keeps showing the correct URL (including
HTTPS) and privacy shield.
- Opening privacy dashboard should always show that protections are
active and that the site is secure.

You can also try with `abc.com`, `cnn.com` (which always should show
`edition.cnn.com`).