Remove onboarding design experiment logic from Launch/Onboarding (#7174)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1207908166761516/task/1212011215258009?focus=true

### Description
This pull request removes the onboarding design experiment logic from
both the launch and onboarding flows. The main impact is that the app no
longer waits for privacy config or uses experimental onboarding page
layouts based on enrollment status. The code is simplified and
dependencies related to the onboarding experiment are removed. The rest
of the onboarding experiment code will be replaced in a follow up PR.

**Removal of onboarding design experiment logic:**

* Eliminated the dependency on `OnboardingDesignExperimentManager` from
the constructors of both `LaunchViewModel` and `OnboardingViewModel`,
simplifying their initialization.
[[1]](diffhunk://#diff-4181f06c936b6bec88667a88ae677aaada2e85c971639f4f06d3989ee3b6db7aL45-R45)
[[2]](diffhunk://#diff-3e7dc4b2bc3ca28e86e1049525ad45efb366b8e0575787f659812b27a3cb4d73L42-L63)
* Removed logic from `LaunchViewModel.determineViewToShow` that waited
for privacy config and referrer data based on the experiment, now always
waiting only for referrer data.
* Deleted experiment-specific onboarding page initialization in
`OnboardingViewModel.initializePages`, now always using the default page
layout manager.

_App Launch/Onboarding_
- [ ] Clean install app
- [ ] Launch
- [ ] Onboarding should launch

### UI changes
N/A